restructure-auth
