+++ src/components/pages/Authentication/components/AvatarBuy.tsx
