+++ src/components/pages/Authentication/components/AvatarCarousel.tsx
