+++ src/components/pages/Authentication/components/CandyDrop.tsx
