+++ src/components/pages/Authentication/components/LottiePlayer.tsx
